moved comment form into different component
